Finished render record.
