+++ b3lb/loadbalancer/settings.py
@@ -249,11 +249,7 @@ B3LB_RECORD_PATH_HIERARCHY_DEPHT = env.int('B3LB_RECORD_PATH_HIERARCHY_DEPHT', d
 ######
 # B3LB Celery Settings
 ######
-
-
-B3LB_SCHEDULE_TASK_CHECK_NODES_QUEUE = env.str("B3LB_SCHEDULE_TASK_CHECK_NODES_QUEUE", default="b3lb")
-B3LB_SCHEDULE_TASK_CLEANUP_ASSETS_QUEUE = env.str("B3LB_SCHEDULE_TASK_CLEANUP_ASSETS_QUEUE", default="b3lb")
-B3LB_SCHEDULE_TASK_UPDATE_SECRET_LISTS_QUEUE = env.str("B3LB_SCHEDULE_TASK_UPDATE_SECRET_LISTS_QUEUE", default="b3lb")
-B3LB_SCHEDULE_TASK_UPDATE_STATISTICS_QUEUE = env.str("B3LB_SCHEDULE_TASK_UPDATE_STATISTICS_QUEUE", default="b3lb")
-B3LB_RECORD_TASK_QUEUE = env.str("B3LB_RECORD_TASK_QUEUE", default="b3lb")
 B3LB_RECORD_TASK_TEMPLATE_FOLDER = env.path("B3LB_RECORD_TASK_TEMPLATE_FOLDER", default="/templates").root
+B3LB_TASK_QUEUE_CORE = env.str("B3LB_TASK_QUEUE_CORE", default="b3lb")
+B3LB_TASK_QUEUE_RECORD = env.str("B3LB_TASK_QUEUE_RECORD", default="b3lb")
+B3LB_TASK_QUEUE_STATISTICS = env.str("B3LB_TASK_QUEUE_STATISTICS", default="b3lb")
